Add methods to convert the response struct to an alist
cl-structs cannot be encoded into JSON with json.el[1] - it raises an error.
This means the JSON-RPC response cannot be sent back to the client, because it
cannot be encoded.

It is nice to be able to manage the response object by encapsulating it within a
cl-struct. To allow this, add helper methods that turn the response (and any
erros) into an alist - these can be used as intermediaries to cast responses
into JSON.

[1] This is as of 22/05/2019. System info:
    - OS: Arch Linux
    - Emacs: 26.2 (built from Pacman, up-to-date)
    - json.el: 1.4