Change `jrpc-nil-p` to check for all null values - nil and json-null
When decoding json, we might well end up with json-null values instead of nil
values. The purpose of this method is to check for null values instead of just
falsey values - since it's going to be used on decoded json, allow json-null as
well.

Also rename the method to indicate this change.