Extract `id` as early as possible so it can be sent back
When responding, the JSON-RPC protocol demands the ID be returned whenever
possible, in order to synchronize messages. To facilitate this, extract the ID
upfront, before the rest of the JSON is checked properly. Ignore any errors
extracting it. This gives us the maximum chance of extracting the ID, thus the
maximum chance it will be returned in an error.