Add integration test for a successful remote call to `+`
Temporarily expose the `+` function and call it. This test should complete
successfully, receive the correct result, and the response should correctly
adhere to the JSON-RPC 2.0 standard.

Currently, this test does not pass - the handler's response does not properly
adhere to JSON-RPC 2.0. It returns an empty `error' trait, rather than none at
all.