Add ability to process batch requests
The JSON-RPC 2.0 specification requires that requests can be sent as lists as
well as individual requests. Decode requests as they come in, determine whether
they are lists or not, then process them.

Note that this does not add tests for batch requests, and does break one
existing test because it changes the way requests are decoded. Requests are now
verified separately to the decoding process. This architecture may be better
overall anyway because it reduces both implicit behaviour and coupling.