Implement transport layer using `web-server.el`
Implement a basic transport layer using HTTP and optional Basic Access
Authentication for security. This layer exposes the RPC protocol over HTTP.

The `web-server` package is also known as `emacs-web-server`.

This implementation is designed to replace the initial implementation with
Elnode. Elnode was not used because the project appears to be dead. It has fatal
bugs that need to be patched to make it usable, and `web-server` seems to be the
standard HTTP server to use within Emacs.

This is a slight shame, because Elnode seems to have a simpler interface for
basic operations, but `web-server` does appear to provide more power.