Add integration test for RPC layer - full call, empty JSON object
The test suite for the `emacs-web-server` transport layer implementation was
failing because of an error in the way the RPC layer's error was being encoded
when supplied with an empty JSON object ("{}"). Add a test to the RPC layer to
narrow down the source of this problem and ensure it isn't coming from the RPC
layer.