Fix duplicate error name
An error with the same name as an error in the RPC layer was being declared in
the transport layer. This was causing its parent class to be overridden, meaning
the RPC layer would no longer catch it properly. Fix this by adding `http` to
all error names in the HTTP layer.