Remove persistent list of exposed functions; pass them as parameter
In order to allow multiple RPC servers to run at once, there cannot be one
central list of exposed functions. Remove this list and instead force the user
to pass the set of legal functions as an extra parameter to `jrpc-handle`.

Another option was to use something like an alist to store multiple function
sets. Instead of this, function exposure is not outsourced to the transport
layer. The transport system can figure out which particular paradigm they want
to use, and implement it. This also allows to the transport layer to swap
between protocols more easily, without having to wrap variables with aliases,
track their states, etc.